Changes again in the page text info plugin tweaks for the widgets affect when enabling flash buttons and menu items: We call UIManager:yieldToEPDC(5000) for all the devices when flashing is enabled in all the commited sources. This fixes the problems mainly for Kobo Clara BW. For the Kindle devices we use double time because it is quicker and for the Kobo Clara Colour a bit more
There are antialiasing problems for buttons highlighted with text using fast refresh for the Kobo Libra Colour and
the Kobo Clara BW devices. We use ui refresh just for these devices

There are problems for button icons highlighted in the fm using fast refresh for Kobo Libra 2, Kobo Clara 2E
and Kindle devices. Part of the file chooser screen is disappears when flashing. We use ui refresh just for the icons
of the fm. But just for these icons.

Kindle devices and the Kobo Libra Colour are so quick when tapping button icons and menu entries in the top menu, that the flash lingers a bit.
We use a little delay for them when updating the top menu to avoid it
It is the same for Android devices, they are even quicker and need more delay